+++ lzma-rust/src/enc/encoder_fast.rs
@@ -11,8 +11,8 @@ impl FashEncoderMode {
     pub const EXTRA_SIZE_BEFORE: u32 = 1;
     pub const EXTRA_SIZE_AFTER: u32 = MATCH_LEN_MAX as u32 - 1;
 
-    pub fn get_memery_usage(dict_size: u32, extra_size_before: u32, mf: MFType) -> u32 {
-        LZEncoder::get_memery_usage(
+    pub fn get_memory_usage(dict_size: u32, extra_size_before: u32, mf: MFType) -> u32 {
+        LZEncoder::get_memory_usage(
             dict_size,
             extra_size_before.max(Self::EXTRA_SIZE_BEFORE),
             Self::EXTRA_SIZE_AFTER,