+++ lzma-rust/src/enc/encoder_normal.rs
@@ -15,8 +15,8 @@ impl NormalEncoderMode {
     const OPTS: u32 = 4096;
     pub const EXTRA_SIZE_BEFORE: u32 = Self::OPTS;
     pub const EXTRA_SIZE_AFTER: u32 = Self::OPTS;
-    pub fn get_memery_usage(dict_size: u32, extra_size_before: u32, mf: MFType) -> u32 {
-        LZEncoder::get_memery_usage(
+    pub fn get_memory_usage(dict_size: u32, extra_size_before: u32, mf: MFType) -> u32 {
+        LZEncoder::get_memory_usage(
             dict_size,
             extra_size_before.max(Self::EXTRA_SIZE_BEFORE),
             Self::EXTRA_SIZE_AFTER,